+++ new-mock-engine-store/src/mock_store/common.rs
@@ -1,0 +1,22 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+pub use collections::{HashMap, HashSet};
+pub use engine_store_ffi::{
+    ffi::{
+        interfaces_ffi,
+        interfaces_ffi::{EngineStoreServerHelper, RaftStoreProxyFFIHelper, RawCppPtr, RawVoidPtr},
+        UnwrapExternCFunc,
+    },
+    TiFlashEngine,
+};
+pub use engine_traits::{
+    Engines, Iterable, KvEngine, Mutable, Peekable, RaftEngine, RaftEngineReadOnly, RaftLogBatch,
+    SyncMutable, WriteBatch, CF_DEFAULT, CF_LOCK, CF_RAFT, CF_WRITE,
+};
+pub use kvproto::{
+    raft_cmdpb::AdminCmdType,
+    raft_serverpb::{PeerState, RaftApplyState, RaftLocalState, RegionLocalState},
+};
+pub use protobuf::Message;
+pub use tikv_util::{box_err, box_try, debug, error, info, warn};
+
+pub use crate::config::MockConfig;
