+++ new-mock-engine-store/src/mock_store/mock_fast_add_peer.rs
@@ -1,0 +1,40 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+
+// use std::{
+//     cell::RefCell,
+//     collections::BTreeMap,
+//     pin::Pin,
+//     sync::{
+//         atomic::{AtomicU64, Ordering},
+//         Mutex,
+//     },
+//     time::Duration,
+// };
+
+// use assert_type_eq;
+// use collections::{HashMap, HashSet};
+// use engine_store_ffi::ffi::{
+//     interfaces_ffi,
+//     interfaces_ffi::{EngineStoreServerHelper, RaftStoreProxyFFIHelper,
+// RawCppPtr, RawVoidPtr},     UnwrapExternCFunc,
+// };
+// use engine_traits::RaftEngineReadOnly;
+// use engine_traits::{
+//     Engines, Iterable, KvEngine, Mutable, Peekable, RaftEngine, RaftLogBatch,
+// SyncMutable,     WriteBatch, CF_DEFAULT, CF_LOCK, CF_RAFT, CF_WRITE,
+// };
+// use int_enum::IntEnum;
+// use kvproto::{
+//     raft_cmdpb::AdminCmdType,
+//     raft_serverpb::{PeerState, RaftApplyState, RaftLocalState,
+// RegionLocalState}, };
+// use protobuf::Message;
+// use tikv_util::{box_err, box_try, debug, error, info, warn};
+
+// use crate::node::NodeCluster;
+
+// use crate::common::*;
+// use crate::mock_engine_store_server::*;
+// use crate::into_engine_store_server_wrap;
+
+// use crate::mock_cluster;
