+++ new-mock-engine-store/src/mock_store/mod.rs
@@ -1,0 +1,10 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+pub(crate) mod common;
+pub mod mock_core;
+pub use mock_core::*;
+pub mod mock_fast_add_peer;
+pub use mock_fast_add_peer::*;
+pub mod mock_engine_store_server;
+pub use mock_engine_store_server::*;
+pub mod mock_page_storage;
+pub use mock_page_storage::*;
